refacotr : nickname modal api 연결 부분 뷰로직 분리
